Refactor helpers.verifyRequestIp.
