+++ lib/helpers.js
@@ -10,6 +10,7 @@ const {
   keystores,
   keyDescriptionStorage
 } = require('bedrock-kms');
+const forwarded = require('forwarded');
 const jsigs = require('jsonld-signatures');
 const {
   extendContextLoader,
@@ -19,6 +20,7 @@ const {
 const {config, util: {BedrockError}} = bedrock;
 const {verifyCapabilityInvocation} = require('http-signature-zcap-verify');
 const {CapabilityDelegation} = require('ocapld');
+const {Netmask} = require('netmask');
 
 // TODO: a lot of helper code here is in common w/ bedrock-data-hub-storage,
 // consider refactoring to share it
@@ -222,6 +224,30 @@ exports.verifyDelegation = async ({keystoreId, host, capability}) => {
     throw error;
   }
   return results;
+};
+
+exports.verifyRequestIp = ({keystoreConfig, req}) => {
+  const {ipAllowList} = keystoreConfig;
+  if(!ipAllowList) {
+    return {verified: true};
+  }
+
+  // the first IP in the sourceAddresses array will *always* be the IP
+  // reported by Express.js via `req.connection.remoteAddress`. Any additional
+  // IPs will be from the `x-forwarded-for` header.
+  const sourceAddresses = forwarded(req);
+
+  // ipAllowList is an array of CIDRs
+  for(const cidr of ipAllowList) {
+    const netmask = new Netmask(cidr);
+    for(const address of sourceAddresses) {
+      if(netmask.contains(address)) {
+        return {verified: true};
+      }
+    }
+  }
+
+  return {verified: false};
 };
 
 /**
