+++ lib/http.js
@@ -16,10 +16,8 @@ const {createMiddleware} = require('bedrock-passport');
 const brZCapStorage = require('bedrock-zcap-storage');
 const cors = require('cors');
 const database = require('bedrock-mongodb');
-const forwarded = require('forwarded');
 const {generateRandom} = require('webkms-switch');
 const helpers = require('./helpers');
-const {Netmask} = require('netmask');
 const {validator: validate} = require('./validator');
 const {
   getKeystoreQuery,
@@ -117,7 +115,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       // retrieval of full configs
 
       const filteredResult = results.filter(({config: keystoreConfig}) => {
-        const {verified} = _verifyRequestIp({keystoreConfig, req});
+        const {verified} = helpers.verifyRequestIp({keystoreConfig, req});
         return verified;
       });
 
@@ -134,7 +132,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
         {req, localId: req.params.keystoreId, routes});
 
       const {config: keystoreConfig} = await keystores.get({id: keystoreId});
-      const {verified} = _verifyRequestIp({keystoreConfig, req});
+      const {verified} = helpers.verifyRequestIp({keystoreConfig, req});
       if(!verified) {
         throw new BedrockError(
           'Permission denied. Source IP is not allowed.', 'NotAllowedError', {
@@ -164,7 +162,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
 
       const {config: keystoreConfig} = await keystores.get({id});
 
-      const {verified} = _verifyRequestIp({keystoreConfig, req});
+      const {verified} = helpers.verifyRequestIp({keystoreConfig, req});
       if(!verified) {
         throw new BedrockError(
           'Permission denied. Source IP is not allowed.', 'NotAllowedError', {
@@ -430,7 +428,7 @@ async function _handleOperation({req, expectedRootCapability, keystoreId}) {
   }
 
   const {config: keystoreConfig} = await keystores.get({id: keystoreId});
-  const {verified} = _verifyRequestIp({keystoreConfig, req});
+  const {verified} = helpers.verifyRequestIp({keystoreConfig, req});
   if(!verified) {
     throw new BedrockError(
       'Permission denied. Source IP is not allowed.', 'NotAllowedError', {
@@ -455,30 +453,6 @@ async function _handleOperation({req, expectedRootCapability, keystoreId}) {
     operation.invocationTarget.id = `${url}/${random}`;
   }
   return runOperation({operation});
-}
-
-function _verifyRequestIp({keystoreConfig, req}) {
-  const {ipAllowList} = keystoreConfig;
-  if(!ipAllowList) {
-    return {verified: true};
-  }
-
-  // the first IP in the sourceAddresses array will *always* be the IP
-  // reported by Express.js via `req.connection.remoteAddress`. Any additional
-  // IPs will be from the `x-forwarded-for` header.
-  const sourceAddresses = forwarded(req);
-
-  // ipAllowList is an array of CIDRs
-  for(const cidr of ipAllowList) {
-    const netmask = new Netmask(cidr);
-    for(const address of sourceAddresses) {
-      if(netmask.contains(address)) {
-        return {verified: true};
-      }
-    }
-  }
-
-  return {verified: false};
 }
 
 // Note: for dereferencing `did:` URLs
